creacion de los migrations de la base de datos

+++ database/migrations/2023_07_18_041713_create_usuario_grupo.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('usuario')->constrained('usuario_grupo');
             $table->foreignId('grupo')->constrained('usuario_grupo');
-            $table->smallInteger('estado');
+            $table->tinyInteger('estado');
             $table->timestamps();
         });
     }
